// src/models/enums/socket-routes.enum.ts
export enum SocketRoutes {
  CS_AUTHENTICATE = 'CS_AUTHENTICATE',
  SC_AUTHENTICATE = 'SC_AUTHENTICATE',

  CS_UPDATE_SETTINGS = 'CS_UPDATE_SETTINGS',
  SC_UPDATE_SETTINGS = 'SC_UPDATE_SETTINGS',
  CS_GET_SETTINGS = 'CS_GET_SETTINGS',
  SC_GET_SETTINGS = 'SC_GET_SETTINGS',

  // NOTIFICATION
  CS_LIST_NOTIFICATION = 'CS_LIST_NOTIFICATION',
  SC_LIST_NOTIFICATION = 'SC_LIST_NOTIFICATION',

  SC_SEND_NOTIFICATION = 'SC_SEND_NOTIFICATION',

  // PAGE
  CS_LIST_PAGE = 'CS_LIST_PAGE',
  SC_LIST_PAGE = 'SC_LIST_PAGE',
  CS_CREATE_PAGE = 'CS_CREATE_PAGE',
  SC_CREATE_PAGE = 'SC_CREATE_PAGE',
  CS_UPDATE_PAGE = 'CS_UPDATE_PAGE',
  SC_UPDATE_PAGE = 'SC_UPDATE_PAGE',
  CS_DELETE_PAGE = 'CS_DELETE_PAGE',
  SC_DELETE_PAGE = 'SC_DELETE_PAGE',

  // COMPONENT
  CS_LIST_COMPONENT = 'CS_LIST_COMPONENT',
  SC_LIST_COMPONENT = 'SC_LIST_COMPONENT',
  CS_CREATE_COMPONENT = 'CS_CREATE_COMPONENT',
  SC_CREATE_COMPONENT = 'SC_CREATE_COMPONENT',
  CS_UPDATE_COMPONENT = 'CS_UPDATE_COMPONENT',
  SC_UPDATE_COMPONENT = 'SC_UPDATE_COMPONENT',
  CS_DELETE_COMPONENT = 'CS_DELETE_COMPONENT',
  SC_DELETE_COMPONENT = 'SC_DELETE_COMPONENT',

  // FILE
  CS_LIST_FILE = 'CS_LIST_FILE',
  SC_LIST_FILE = 'SC_LIST_FILE',
  CS_CREATE_FILE = 'CS_CREATE_FILE',
  SC_CREATE_FILE = 'SC_CREATE_FILE',
  CS_UPDATE_FILE = 'CS_UPDATE_FILE',
  SC_UPDATE_FILE = 'SC_UPDATE_FILE',
  CS_DELETE_FILE = 'CS_DELETE_FILE',
  SC_DELETE_FILE = 'SC_DELETE_FILE',
}